+++ agent/consul/discovery_chain_endpoint_test.go
@@ -310,7 +310,7 @@ func TestDiscoveryChainEndpoint_Get_BlockOnNoChange(t *testing.T) {
 		})
 
 		g.Go(func() error {
-			for i := uint64(0); i < 200; i++ {
+			for i := 0; i < 200; i++ {
 				time.Sleep(5 * time.Millisecond)
 
 				var out bool
